Move `PubRandCommitError` and `FinalitySigError` into error.rs
